+++ src/main/java/de/ellpeck/actuallyadditions/data/BlockStateGenerator.java
@@ -7,14 +7,9 @@ import net.minecraft.core.Direction;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.data.PackOutput;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.CropBlock;
-import net.minecraft.world.level.block.SlabBlock;
-import net.minecraft.world.level.block.StairBlock;
-import net.minecraft.world.level.block.WallBlock;
+import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.neoforged.neoforge.client.model.generators.BlockStateProvider;
-import net.neoforged.neoforge.client.model.generators.ConfiguredModel;
 import net.neoforged.neoforge.client.model.generators.ModelFile;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
 
@@ -227,13 +222,17 @@ public class BlockStateGenerator extends BlockStateProvider {
 
     private void buildLitState(Supplier<Block> block) {
         ResourceLocation name = BuiltInRegistries.BLOCK.getKey(block.get());
-        assert name != null;
+        ModelFile offModel = models().withExistingParent(name.toString(), "cube_all")
+            .texture("all", modLoc("block/" + name.getPath()));
+        ModelFile onModel = models().withExistingParent(name.toString() + "_on", "cube_all")
+            .texture("all", modLoc("block/" + name.getPath() + "_on"));
+
 
         getVariantBuilder(block.get())
             .partialState().with(BlockStateProperties.LIT, false)
-            .addModels(ConfiguredModel.builder().modelFile(models().cubeAll(name.toString(), modLoc("block/" + name.getPath()))).build())
+            .modelForState().modelFile(offModel).addModel()
             .partialState().with(BlockStateProperties.LIT, true)
-            .addModels(ConfiguredModel.builder().modelFile(models().cubeAll(name.toString(), modLoc("block/" + name.getPath() + "_on"))).build());
+            .modelForState().modelFile(onModel).addModel();
     }
 
     private void fullyDirectionalBlock(Supplier<Block> block) {