+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 * Fixed the Sack not saving filter options, and Auto-Insert state always being stuck on.
 * Fixed no item pickup sound when the Sack picks things up.
 * Fixed Advanced Item Laser Relays filters not working, and the Smart filter buttons not working.
+* Added recipes for the Lamps.
+* Added recipe for the Lamp Controller.
+* Fixed Lamp Controller model rotation.
+* Fixed Lamp textures being always lit.
 
 # 1.3.12+mc1.21.1
 * Fixed the Drill dye recipes not functioning.